+++ R/stat-poly-line.R
@@ -100,10 +100,24 @@
 #'   rows of predicted values and their confidence limits. Optionally it will
 #'   also include additional values related to the model fit.
 #'
+#' @section Model fit methods supported: Several model fit functions are
+#'   supported explicitly, and some of their differences smoothed out. The
+#'   compatibility is checked late, based on the class of the returned fitted
+#'   model object. This makes it possible to use wrapper functions that do model
+#'   selection or other adjustments to the fit procedure on a per panel or per
+#'   group basis. In the case of fitted model objects of classes not explicitly
+#'   supported an attempt is made to find the usual accessors, and if found,
+#'   either complete or partial support frequently just works. The argument to
+#'   parameter \code{method} can be either the name of a function or a character
+#'   string giving the name. This approach makes it possible to support model
+#'   fit functions that are not dependencies of 'ggpmisc'. Either attach the
+#'   package where the function is defined and pass it by name or as string, or
+#'   use colon notation when passing the name of the function.
+#'
 #' @section Computed variables: `stat_poly_line()` provides the following
-#'   variables, some of which depend on the orientation: \describe{ \item{y *or*
-#'   x}{predicted value} \item{ymin *or* xmin}{lower pointwise confidence
-#'   interval around the mean} \item{ymax *or* xmax}{upper pointwise confidence
+#'   variables, some of which depend on the orientation: \describe{ \item{y \strong{or}
+#'   x}{predicted value} \item{ymin \strong{or} xmin}{lower pointwise confidence
+#'   interval around the mean} \item{ymax \strong{or} xmax}{upper pointwise confidence
 #'   interval around the mean} \item{se}{standard error} }
 #'
 #'   If \code{fm.values = TRUE} is passed then columns based on the summary of
@@ -120,9 +134,11 @@
 #'   (\code{"geom_smooth"} is the default) are understood and grouping
 #'   respected.
 #'
-#' @family ggplot statistics for linear and polynomial regression
+#' @note Currently confidence bands for the regression band are not plotted in
+#'   some cases, and in the case of MA and SMA model, it only displays the
+#'   uncertainty of the slope.
 #'
-#' @export
+#' @family ggplot statistics for linear and polynomial regression
 #'
 #' @examples
 #' ggplot(mpg, aes(displ, hwy)) +
@@ -407,9 +423,18 @@ poly_line_compute_group_fun <-
     } else {
       if (class(fm)[1] == "sma") {
         if (se) {
-          b0 <- unlist(fm$coef[[1]]["elevation", ])
+          message("SMA/MA, band is currently for slope only!")
+          # fm$coef[[1]] is a data.frame
+          #
+          #           coef(SMA) lower limit upper limit
+          # elevation 39.441685   38.011038    40.87233
+          # slope     -4.609003   -5.008148    -4.24167
+          ### parameter estimates assumed independent!!
+          ## bootstraping would be more appropriate
+          coef.sma <- fm$coef[[1]]
+          b0 <- unlist(coef.sma["elevation", ])
           b0 <- ifelse(is.na(b0), 0, b0)
-          b1 <- unlist(fm$coef[[1]]["slope", ])
+          b1 <- unlist(coef.sma["slope", ])
           # centering is needed
           b0delta <- b0 - b0[1]
           if (all(b0 == 0)) {
@@ -417,7 +442,7 @@ poly_line_compute_group_fun <-
             center.y <- 0
             center.x <- 0
           } else {
-            # center on data means
+            # data centroid
             center.y <- mean(data[["y"]])
             center.x <- mean(data[["x"]])
           }
@@ -426,14 +451,18 @@ poly_line_compute_group_fun <-
           prediction <- data.frame(x = xseq,
                                    y = center.y + b0delta[1] +
                                      b1[1] * (xseq - center.x),
-                                   ymin = center.y + b0delta[2] +
+                                   ymin = center.y +
                                      (rightside * b1[2] + leftside * b1[3]) *
                                      (xseq - center.x),
-                                   ymax = center.y + b0delta[3] +
+                                   ymax = center.y +
                                      (rightside * b1[3] + leftside * b1[2]) *
                                      (xseq - center.x))
-        } else {
+        } else { # se is FALSE
           coefs <- stats::coefficients(fm)
+          # named vector
+          #
+          # elevation     slope
+          # 39.441685 -4.609003
           prediction <-
             data.frame(x = xseq, # data[["x"]]
                        y = coefs["elevation"] + coefs["slope"] * xseq)